Feature: Added error components for each page, suspense for each page separately
Feature: I made it so that any error would limit only a block of the page, not the entire page.

Infra: improved the button and text components. The button can now receive mouse events. You can no longer pass type='p' to the text, because it is the default.

Important: the code is currently displaying an error. The code doesn't actually work.